requirements.txt
Django==3.0.4
django-octicons==1.0.2
django-markdownx==3.0.1